Complete Production-Ready Features and Installer Fixes

+++ Vantus.Core/Models/EngineStatus.cs
@@ -2,7 +2,7 @@ namespace Vantus.Core.Models;
 
 public class EngineStatus
 {
-    public string State { get; set; } = "Unknown";
-    public long IndexedCount { get; set; }
+    public string State { get; set; } = "Unknown"; // e.g., "Idle", "Indexing", "Paused"
+    public int IndexedCount { get; set; }
     public bool IsCrawling { get; set; }
 }
